requirements.txt
tensorflow
numpy
pillow
matplotlib
colorama
keras
rich
Levenshtein
